+++ c10/xpu/XPUCachingAllocator.h
@@ -122,32 +122,7 @@ C10_XPU_API int getPoolUseCount(
 } // namespace c10::xpu::XPUCachingAllocator
 
 namespace c10::xpu {
-
+// Keep BC only
 using c10::CaptureId_t;
 using c10::MempoolId_t;
-struct C10_XPU_API MemPool {
-  MemPool(
-      XPUCachingAllocator::XPUAllocator* allocator = nullptr,
-      bool is_user_created = true,
-      bool use_on_oom = false);
-  MemPool(const MemPool&) = delete;
-  MemPool(MemPool&&) = default;
-  MemPool& operator=(const MemPool&) = delete;
-  MemPool& operator=(MemPool&&) = default;
-  ~MemPool();
-
-  MempoolId_t id();
-  XPUCachingAllocator::XPUAllocator* allocator();
-  int use_count();
-  c10::DeviceIndex device();
-  static MempoolId_t graph_pool_handle(bool is_user_created = true);
-
- private:
-  static std::atomic<CaptureId_t> uid_;
-  static std::atomic<CaptureId_t> uuid_;
-  XPUCachingAllocator::XPUAllocator* allocator_;
-  bool is_user_created_;
-  MempoolId_t id_;
-  c10::DeviceIndex device_;
-};
 } // namespace c10::xpu
